fix
max_lifetime: doc.dbconf.max_lifetime * time.minute idle_timeout:
doc.dbconf.idle_timeout * time.minute get_timeout:
doc.dbconf.get_timeout * time.second

+++ backend/internal/middleware/config_loader/config_struct.v
@@ -41,9 +41,9 @@ pub:
 	// 连接池配置
 	max_conns      int = 100 // 默认 100 个
 	min_idle_conns int = 10  // 默认 10个
-	max_lifetime   int = 60  // 默认 60 minute
-	idle_timeout   int = 30  // 默认 30 minute
-	get_timeout    int = 3   // 默认 3 second
+	max_lifetime   i64 = 60  // 默认 60 minute
+	idle_timeout   i64 = 30  // 默认 30 minute
+	get_timeout    i64 = 3   // 默认 3 second
 }
 
 // // MiddlewaresConf is the config of middlewares.
